ZG-133 服务端增加gm命令

+++ services/game/hero/options.go
@@ -57,7 +57,7 @@ func Exp(exp int32) Option {
 	}
 }
 
-func Level(level int8) Option {
+func Level(level int16) Option {
 	return func(o *Options) {
 		o.Level = level
 	}
